Fix zoning update invocation and document zoning guidelines
